+++ service/elasticache/api_op_ModifyUser.go
@@ -43,7 +43,7 @@ type ModifyUserInput struct {
 	// Specifies how to authenticate the user.
 	AuthenticationMode *types.AuthenticationMode
 
-	// The engine for a specific user.
+	// Modifies the engine listed for a user. The options are valkey or redis.
 	Engine *string
 
 	// Indicates no password is required for the user.
@@ -66,7 +66,7 @@ type ModifyUserOutput struct {
 	// Denotes whether the user requires a password to authenticate.
 	Authentication *types.Authentication
 
-	// The current supported value is Redis.
+	// The options are valkey or redis.
 	Engine *string
 
 	// The minimum engine version required, which is Redis OSS 6.0